Fixed some issues
The staff module is making a call to api.quotable.io which has suddenly gone missing.  In case it comes back, for now I'm just testing and using a default answer of REDACTED/Anonymous, which is a good idea in case it fails in the future.  Also added a bit of useful styling to the fortune module to make it read a bit cleaner.